Fix tests; move back shipping transport utils to previous module
