+++ saleor/graphql/checkout/tests/benchmark/test_checkout_mutations.py
@@ -566,7 +566,7 @@ def test_create_checkout_with_order_promotion(
     }
 
     # when
-    with django_assert_num_queries(79):
+    with django_assert_num_queries(80):
         response = user_api_client.post_graphql(MUTATION_CHECKOUT_CREATE, variables)
 
     # then
